06-generative-ai/02-Build-AI-Agent/requirements.txt
azure-identity
azure-ai-projects
azure-ai-agents>=1.0.0b1
python-dotenv
pandas
matplotlib